requirements_gpu.txt
numpy
pandas==1.1.5
mlflow==1.22.0
pylint==2.6.0
pytest-cov==2.10.1
pyyaml==5.4.1
python-json-logger==2.0.2
polyaxon==1.11.3
google-cloud-storage==1.43.0
hydra-core==1.1.1
hydra-optuna-sweeper==1.1.1
optuna==2.10.0
sphinx==4.3.2
jsonlines==3.0.0
seaborn==0.11.2
opencv-python==4.1.2.30
opencv-python-headless==4.1.2.30
scikit-learn
scikit-image
Pillow==8.3.2
Keras==2.2.5
tensorflow==2.2.0
h5py==2.10.0
progressbar2
tqdm==4.45.0
pconst==1.0.0
cython==0.29.23
shapely==1.8.0